+++ packages/app-store/routing-forms/lib/raqbUtils.ts
@@ -5,14 +5,14 @@ import logger from "@calcom/lib/logger";
 import { safeStringify } from "@calcom/lib/safeStringify";
 import { AttributeType } from "@calcom/prisma/enums";
 
-import type { AttributesQueryBuilderConfigWithRaqbFields } from "../lib/getQueryBuilderConfig";
-import { getQueryBuilderConfigForAttributes } from "../lib/getQueryBuilderConfig";
+import type { AttributesQueryBuilderConfigWithRaqbFields } from "./getQueryBuilderConfig";
+import { getQueryBuilderConfigForAttributes } from "./getQueryBuilderConfig";
 import type { Attribute, AttributesQueryValue } from "../types/types";
 import type { LocalRoute } from "../types/types";
 import type { FormResponse, SerializableForm } from "../types/types";
 import type { SerializableField } from "../types/types";
 
-const moduleLogger = logger.getSubLogger({ prefix: ["routing-forms/trpc/raqbUtils"] });
+const moduleLogger = logger.getSubLogger({ prefix: ["routing-forms/lib/raqbUtils"] });
 
 type GetFieldResponse = ({
   field,
